Removing SnapshotArchivePackage
